+++ src/main/java/com/github/zastrixarundell/toramsensei/objects/toram/monsters/LevelingMonster.java
@@ -16,33 +16,34 @@
  *
  */
 
-package com.github.zastrixarundell.toramsensei.objects.toram;
+package com.github.zastrixarundell.toramsensei.objects.toram.monsters;
 
-import com.github.zastrixarundell.toramsensei.Parser;
 import org.jsoup.nodes.Element;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class NPC
+public class LevelingMonster
 {
 
     private String name, level, location, link;
     private List<String[]> exp = new ArrayList<>();
 
-    public NPC(Element npcData)
+    public LevelingMonster(Element npcData)
     {
         //Name, type and duration
-        level = npcData.getElementsByTag("td").first().ownText();
-        location = npcData.getElementsByTag("td").get(1).ownText();
-        link = "http://coryn.club/" + npcData.getElementsByTag("td").get(1)
-                .getElementsByTag("a").first()
+        level = npcData.getElementsByClass("level-col-1").first().text().split(" ")[1];
+        //level = npcData.getElementsByTag("td").first().ownText();
+
+        Element nameAndLocationCol = npcData.getElementsByClass("level-col-2").first();
+
+        name = nameAndLocationCol.getElementsByTag("p").first().text();
+        location = nameAndLocationCol.getElementsByTag("p").last().text();
+
+        link = "http://coryn.club/" + npcData.getElementsByTag("a").first()
                 .attr("href");
 
-        name = Parser.nameParser(npcData.getElementsByTag("td").get(1)
-                .getElementsByTag("a").first().ownText());
-
-        String stringExp = npcData.getElementsByTag("td").last().text();
+        String stringExp = npcData.getElementsByClass("level-col-3").first().text();
         stringExp = stringExp.replaceAll("%", "\t\n");
         stringExp = stringExp.replaceAll("\t", "%");